clean up code for pinboard script
